src/data_preprocessing/01_rcc_metadata/requirements.txt
pandas>=2.1
lxml>=4.9
openslide-python>=1.3.1
tifffile>=2023.2.3
